requirements.txt
asgiref
beautifulsoup4
certifi
charset-normalizer
crispy-bootstrap5
defusedxml
Django
django-bootstrap-v5
django-crispy-forms
django-extensions
future
idna
jira
keyring
ldap3
oauthlib
pyad
pyasn1
requests
requests-oauthlib
requests-toolbelt
soupsieve
sqlparse
typing_extensions
tzdata
urllib3
gunicorn
